clean up and separate out swaps
